Format, get rid of conversion type warning.

+++ ctl-lib/ctl-action.c
@@ -233,7 +233,7 @@ PUBLIC CtlActionT *ActionConfig(AFB_ApiT apiHandle, json_object *actionsJ, int e
 
     // action array is close with a nullvalue;
     if (json_object_get_type(actionsJ) == json_type_array) {
-        int count = json_object_array_length(actionsJ);
+        size_t count = json_object_array_length(actionsJ);
         actions = calloc(count + 1, sizeof (CtlActionT));
 
         for (int idx = 0; idx < count; idx++) {
